+++ src/services/Users.php
@@ -10,80 +10,71 @@
 
 namespace anubarak\seeder\services;
 
-use craft\elements\Asset;
-use craft\elements\Category;
-use craft\elements\Entry;
+use craft\base\Element;
 use craft\elements\User;
-use craft\errors\FieldNotFoundException;
-use craft\helpers\Console;
-use Faker\Factory;
-use Faker\Provider\Person;
 use anubarak\seeder\Seeder;
 
 use Craft;
 use craft\base\Component;
-use yii\base\Model;
 
 /**
- * Users Service
+ * Class Users
  *
- * All of your plugin’s business logic should go in services, including saving data,
- * retrieving data, etc. They provide APIs that your controllers, template variables,
- * and other plugins can interact with.
- *
- * https://craftcms.com/docs/plugins/services
- *
- * @author    Studio Espresso
- * @package   Seeder
- * @since     1.0.0
+ * @package anubarak\seeder\services
+ * @since   26.06.2024
+ * @author  by Robin Schambach
  */
 class Users extends Component
 {
     /**
-     * @param null $group
-     * @param int  $count
+     * @param \craft\models\UserGroup[] $userGroups
      *
-     * @return bool
      * @throws \craft\errors\ElementNotFoundException
      * @throws \yii\base\Exception
      * @throws \Throwable
      */
-    public function generate($group = null, $count = 20): bool
+    public function generate(array $userGroups = [], int $count = 20, callable $cb = null): bool
     {
-        $userGroup = null;
-        if($group !== null){
-            if (ctype_digit($group)) {
-                $userGroup = Craft::$app->userGroups->getGroupById((int)$group);
-            } else {
-                $userGroup = Craft::$app->userGroups->getGroupByHandle($group);
-            }
+        $seeder = Seeder::$plugin->getSeeder();
+        $faker = $seeder->factory;
+
+        $elements = Craft::$app->getElements();
+        $users = Craft::$app->getUsers();
+
+        $userGroupIds = [];
+        foreach ($userGroups as $group) {
+            $userGroupIds[] = $group->id;
         }
 
-        $faker = Factory::create();
+        $transaction = Craft::$app->getDb()->beginTransaction();
+        try {
+            for ($x = 1; $x <= $count; $x++) {
+                $user = new User();
+                $user->passwordResetRequired = true;
+                $user->email = $faker->email();
+                $user->username = $user->email;
+                $user->firstName = $faker->firstName();
+                $user->lastName = $faker->lastName();
+                $elements->saveElement($user);
+                $seeder->saveSeededUser($user);
+                $user->setScenario(Element::SCENARIO_LIVE);
+                $seeder->populateFields($user);
+                $elements->saveElement($user);
 
-        $fields = Craft::$app->fields->getFieldsByElementType('craft\elements\User');
-        $current = 0;
-        Console::startProgress($current, $count);
-        for ($x = 1; $x <= $count; $x++) {
-            $user = new User();
-            $user->passwordResetRequired = true;
-            $user->email = $faker->email;
-            $user->username = $user->email;
-            $user->firstName = $faker->firstName;
-            $user->lastName = $faker->lastName;
-            Craft::$app->elements->saveElement($user);
-            Seeder::$plugin->seeder->saveSeededUser($user);
-            Seeder::$plugin->seeder->populateFields($fields, $user);
-            Craft::$app->elements->saveElement($user);
-            if($userGroup !== null){
-                Craft::$app->users->assignUserToGroups($user->id, [$userGroup->id]);
+                if (!empty($userGroupIds)) {
+                    $users->assignUserToGroups($user->id, $userGroupIds);
+                }
+                if ($cb) {
+                    $cb($x, $count);
+                }
             }
-            $current++;
-            Console::updateProgress($current, $count);
+
+            $transaction->commit();
+        } catch (\Throwable $throwable) {
+            $transaction->rollBack();
+            throw $throwable;
         }
-        Console::endProgress();
 
         return true;
     }
-
 }